Final touches

+++ cross-compiler/build.sh
@@ -3,15 +3,17 @@
 # handle arguments
 #deploy=$(echo "$@" | awk -F= '{a[$1]=$2} END {print(a["--deploy"])}')
 # key-value pairs (HOW IS THERE NOT A BUILT IN FOR THIS?)
+processes=8
 while test $# != 0
 do
     left=$(echo "$1" | cut -d "=" -f 1)
     right=$(echo "$1" | cut -d "=" -f 2)
     case "$left" in
-    --skip-rsync) skip_rsync=t ;;
+    --sync) sync=t ;;
     --clean-build) clean_build=t ;;
     --debug) debug=t ;;
     --deploy) deploy=$right ;;
+    --processes) processes=$right ;;
     esac
     shift
 done
@@ -20,7 +22,7 @@ done
 # check for mounts
 if ! mountpoint -q $SYSROOT
 then
-    echo "Rootfs not mounted. Run container with -v rootfs:$SYSROOT"
+    echo "Rootfs not mounted. Run container with -v [local_fs|live_fs]:$SYSROOT"
     exit 1;
 fi
 
@@ -34,7 +36,7 @@ fi
 # rsync options:
 # verbose, recursive through directories, keep symlink, Relative,
 # delete missing, discard links outside of transfered directory
-if ! [ "$skip_rsync" == t ]
+if [ "$sync" == t ]
 then
     # reason for --copy-unsafe-links: 
     # rsync has no --convert-unsafe-absolute-to-relative option or something
@@ -54,14 +56,11 @@ then
     rsync -tr --delete-after --links --copy-unsafe-links --ignore-errors \
         --rsh "/usr/bin/sshpass -p $REMOTE_PASSWORD ssh -o StrictHostKeyChecking=no -l $REMOTE_USER" \
         --rsync-path="sudo rsync" --timeout=3 \
-        --include='/' \
-        --include='/lib/' \
-        --include='/lib/***' \
-        --include='/usr/' \
-        --include='/usr/***' \
-        --exclude='*' \
-        $REMOTE_USER@$REMOTE_IP:/ $SYSROOT \
+        $REMOTE_USER@$REMOTE_IP:/usr $SYSROOT \
         || true # do not fail on error, for isntance, because pi is down
+
+    # little hacky, but okay:
+    ln -sf $SYSROOT/usr/lib $SYSROOT/lib
     echo "Rsync done or failed"
 fi
 
@@ -93,20 +92,23 @@ fi
 
 cmake -DCMAKE_TOOLCHAIN_FILE=$CROSS_TOOLCHAIN $CMAKE_EXTRA ..
 #cmake --build . --parallel 4 # looks cleaner, but somehow broken
-make -j8 # TODO: make the thread count a parameter that can be passed to docker
+make -j $processes # TODO: make the thread count a parameter that can be passed to docker
 
 # TODO: error handling, so we dont upload failed builds
 
-# upload if necessary
-if ! [ -z $deploy ]
+# upload if necessary and if make was a success
+if [ ! -z $deploy ] && [ $? -eq 0 ]
 then
     echo -n "Deploying build... "
     cd /package && \
     rsync -rR --delete-after --links --copy-unsafe-links --perms \
         --rsh "/usr/bin/sshpass -p $REMOTE_PASSWORD ssh -o StrictHostKeyChecking=no -l $REMOTE_USER" \
-        --rsync-path="sudo rsync" --timeout=3 \
+        --timeout=3 \
         build-$GNU_HOST $REMOTE_USER@$REMOTE_IP:"$deploy"
+    #--rsync-path="sudo rsync" \
     echo "build deployed"
+else
+    echo "Not deploying."
 fi
 
 echo "finished"
